Fix tests to cover uncovered branches
This commit solves the coverage drop from:

- mbland/test-page-opener#23
  mbland/test-page-opener@01a79f6

Restores `istanbul` as the coverage reporter for the JSDom-based CI run
to restore consistency in reporting, particularly in Coveralls.

---

Adapted from the new test/missing-app-div/jsdom.test.js:

Moved the "missing #app div" test into a separate directory from the
other tests in ../jsdom.test.js. to prevent Node.js from using the same
`test-modules/main.js` import. Otherwise:

- If the test case were in the same file, the "missing #app div" branch
  wouldn't execute and the test would fail.

- If this test file were in the same directory, the Istanbul coverage
  reporter wouldn't see the coverage from the "missing #app div" branch.
  I don't know exactly why that is.

At the same time, the previous `src="./main.js?version=missing"` query
suffix is no longer necessary.

I got the idea to organize the tests this way after successfully
covering similar code in:

- mbland/tomcat-servlet-testing-example#85
  mbland/tomcat-servlet-testing-example@b5df30e
