+++ test-modules/missing.html
@@ -7,10 +7,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Test Page for JsdomPageOpener missing an app div</title>
-    <!-- Provide a query suffix so import from index.html doesn't get reused and
-       - cause "JsdomPageOpener > doesn't throw if missing app div" to fail.
-       - https://github.com/nodejs/help/issues/2751#issuecomment-1075535742 -->
-    <script type="module" src="./main.js?version=missing"></script>
+    <script type="module" src="./main.js"></script>
   </head>
   <body>
     <div id="not-the-div-you're-looking-for"></div>
