+++ test/jsdom-fixture.js
@@ -1,0 +1,15 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+export default class JsdomFixture {
+  static #origWindow = globalThis.window
+  static #origDocument = globalThis.document
+
+  restoreOrigWindowAndDocument() {
+    globalThis.window = JsdomFixture.#origWindow
+    globalThis.document = JsdomFixture.#origDocument
+  }
+}
